requirements.txt
annotated-types==0.7.0
greenlet==3.3.1
iniconfig==2.3.0
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pytest==9.0.2
PyYAML==6.0.3
selectolax==0.4.6
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
